# docs/gatekeeper/openapi3_0.yaml
openapi: 3.0.4
info:
  title: Gatekeeper
  description: auth services for codeforge to handle permissions and auth between and for components.
  version: 0.0.1

servers:
  - url: http://api.example.com/gatekeeper
    description: placeholder

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: X-API-KEY

paths:
  /user/{id}:
    get:
      summary: Returns the user matching the id.
      parameters:
        - in: path
          name: id # Note the name is the same as in the path
          required: true
          schema:
            type: string
            minimum: 1
          description: The user ID
      security:
        - CookieAuth: []
      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: delete selected user
      
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: The user ID
      security:
        - CookieAuth: []
      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
                properties: 
                  reason:
                    type: string
                example:
                  reason: user deleted
    put:
      summary: updates the user matching the id.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: The user ID
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
                org_id:
                  type: string
                team_id:
                  type: string
                role_id:
                  type: string

              example: # Sample object
                username: test
                password: test_password
                email: test@test.com
                team_id: 6392b7f9-7ed9-430d-8e57-2eedb33df463
                org_id: 74c581fb-2ec1-4252-9901-ef7887eebd05
                role_id: e3f9874d-b578-483a-8e56-230134d795d1
      security:
        - CookieAuth: []
      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
  /signup:
    post:
      summary: creates a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
              example: # Sample object
                username: test
                password: test_password
                email: test@test.com

      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
                properties: 
                  reason:
                    type: string
                  user_created:
                    type: boolean
                example:
                  reason: user created
                  user_created: true
  /token:
    get:
      summary: creates and returns a bearer token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
              example: # Sample object
                username: test
                password: test_password
                email: test@test.com

      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
                properties: 
                  reason:
                    type: string
                  user_created:
                    type: boolean
                example:
                  reason: user created
                  user_created: true

  /signin:
    get:
      summary: signs in the user and creates sesssion
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
              example: # Sample object
                username: test
                password: test_password
                email: test@test.com

      responses:
        "200": # status code
          description: A JSON of a users data
          content:
            application/json:
              schema:
                type: object
                properties: 
                  reason:
                    type: string
                  user_created:
                    type: boolean
                example:
                  reason: user created
                  user_created: true