# src/gatekeeper/main.go
package main

import (
	"codeforge/src/gatekeeper/routes"
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
)

func main() {
	expectedHost := "127.0.0.1:8081"

	router := gin.Default()

	router.Use(gin.Recovery())

	router.Use(func(c *gin.Context) {
		if c.Request.Host != expectedHost {
			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Invalid host header"})
			return
		}
		c.Header("X-Frame-Options", "DENY")
		c.Header("Content-Security-Policy", "default-src 'self'; connect-src *; font-src *; script-src-elem * 'unsafe-inline'; img-src * data:; style-src * 'unsafe-inline';")
		c.Header("X-XSS-Protection", "1; mode=block")
		c.Header("Strict-Transport-Security", "max-age=31536000; includeSubDomains")
		c.Header("Referrer-Policy", "strict-origin")
		c.Header("X-Content-Type-Options", "nosniff")
		c.Header("Permissions-Policy", "geolocation=(),midi=(),sync-xhr=(),microphone=(),camera=(),magnetometer=(),gyroscope=(),fullscreen=(self),payment=()")
		c.Next()
	})

	router.POST("/signup", routes.PostUser)
	router.GET("/token", routes.GetToken)
	router.GET("/signin", routes.GetSignIn)

	authorized := router.Group("/")
	authorized.PUT("/user/:id", routes.PutUser)
	authorized.DELETE("/user/:id", routes.DeleteUser)
	authorized.GET("/user/:id", routes.GetUser)

	err := router.Run("0.0.0.0:8081")
	if err != nil {
		log.Println("failed to start api")
	}
}

// use graphql for the addons using role to restrict data access